Move PlainTarget constructors from TokenInsertionRemovalBehavior.ts to PlainTarget.ts.
